naming threads and processes for logs
