+++ frigate/log.py
@@ -6,7 +6,7 @@ from logging import handlers
 def listener_configurer():
     root = logging.getLogger()
     console_handler = logging.StreamHandler()
-    formatter = logging.Formatter('%(processName)-12s %(threadName)-12s %(name)-16s %(levelname)-8s: %(message)s')
+    formatter = logging.Formatter('%(threadName)-25s %(name)-16s %(levelname)-8s: %(message)s')
     console_handler.setFormatter(formatter)
     root.addHandler(console_handler)
     root.setLevel(logging.INFO)
@@ -18,6 +18,7 @@ def root_configurer(queue):
     root.setLevel(logging.INFO)
 
 def log_process(queue):
+    threading.current_thread().name = f"logger"
     listener_configurer()
     while True:
         record = queue.get()
