+++ frigate/mqtt.py
@@ -1,5 +1,6 @@
 import logging
 import paho.mqtt.client as mqtt
+import threading
 
 from frigate.config import MqttConfig
 
@@ -8,8 +9,7 @@ logger = logging.getLogger(__name__)
 def create_mqtt_client(config: MqttConfig):
     client = mqtt.Client(client_id=config.client_id)
     def on_connect(client, userdata, flags, rc):
-        # TODO: use logging library
-        logger.info("On connect called")
+        threading.current_thread().name = "mqtt"
         if rc != 0:
             if rc == 3:
                 logger.error("MQTT Server unavailable")
@@ -19,6 +19,8 @@ def create_mqtt_client(config: MqttConfig):
                 logger.error("MQTT Not authorized")
             else:
                 logger.error("Unable to connect to MQTT: Connection refused. Error code: " + str(rc))
+            
+        logger.info("MQTT connected")
         client.publish(config.topic_prefix+'/available', 'online', retain=True)       
     client.on_connect = on_connect
     client.will_set(config.topic_prefix+'/available', payload='offline', qos=1, retain=True)
